+++ src/app.js
@@ -14,21 +14,20 @@ import {
   lstorage_cFilter, lstorage_cBookingPlatform, 
   lstorage_cBookingScore
 } from "./utils/appStorage.js";
-import { clearmoduleLocalStorages, reloadModuleCache,setModuleAmagProTimeTestStyle } from "./utils/moduleStorage.js";
+import { clearmoduleLocalStorages, reloadModuleCache,setModuleAmagProTimeTestStyle } from "./utils/modules/moduleStorage.js";
 import { 
   defaultProfileName, consoleWarnMessage_showMessageTurnedOff, 
-  module_details_classVisible, default_e 
+  module_details_classVisible, default_e, buttonTabPreValue,nameInputInvalidRegexep
 } from "./utils/defaults/defaultVariables.js";
 import { profileManager } from "./utils/profileManager.js";
 import { eventListenerHandler, reloadAppAfterChangeHandler } from "./utils/functionHandlers.js";
-import { changeModuleEEWidgetHeightHandler } from "./utils/moduleGlobalUtils.js";
+import { changeModuleEEWidgetHeightHandler } from "./utils/modules/moduleGlobalUtils.js";
 import { generateThemes } from "./components/ui/selectThemes/selectThemes.js";
 import { setScoreValues } from "./utils/setScorevalues.js";
 import { debugStick } from "./utils/appDebugStick.js";
 import { markTabButtons, showHideStatusBar } from "./utils/elementChangers.js";
 import { setStatusBarText } from "./utils/setStatusBarText.js";
-import { nameInputInvalidRegexep } from "./utils/defaults/defaultVariables.js";
-
+import { moduleStorage_preValueConfiguration } from "./utils/modules/defaults/defaultModuleVariables.js";
 // savety function to prevent unwanted webpage content manipulation (triggered by window.onload)
 function isTimeCopy() {
   try {
@@ -92,7 +91,7 @@ document.addEventListener('DOMContentLoaded', async function () {
   const messagesHeadline = document.getElementById('messages-headline')
   const elem_messageSection = document.getElementById('messages-section')
   const configurationsContainer = document.getElementById('config-container')
-  const configWindow_getAll = document.getElementsByClassName('configuration-window')
+  const configWindow_getAll = document.getElementsByClassName('configurationWindow')
   const configProfileName = document.getElementById('configProfileName')
   const profilePictureUser = document.getElementById('profile_picture_user')
   const profileSVG = document.getElementById('Profile')
@@ -103,11 +102,11 @@ document.addEventListener('DOMContentLoaded', async function () {
   const statusBar = document.getElementById('statusBar')
   // tab buttons
   const buttonsTab_getAll = document.getElementsByClassName('button-config-tab')
-  const buttonTab_General = document.querySelector('button#button-tab-general')
-  const buttonTab_Timesheets = document.querySelector('button#button-tab-timesheets')
-  const buttonTab_Bookingsheets = document.querySelector('button#button-tab-bookingsheets')
-  const buttonTab_Projects = document.querySelector('button#button-tab-projects')
-  const buttonBackToMain = document.querySelector('button#buttonBackToMain')
+  const buttonTab_General = document.querySelector('button#'+buttonTabPreValue+'general')
+  const buttonTab_Timesheets = document.querySelector('button#'+buttonTabPreValue+'filters')
+  const buttonTab_Bookingsheets = document.querySelector('button#'+buttonTabPreValue+'platforms')
+  const buttonTab_Projects = document.querySelector('button#'+buttonTabPreValue+'detections')
+  const button_backToMain = document.querySelector('button#button_backToMain')
   // main buttons
   const fillButton = document.querySelector('button#fillButton')
   const fillCancelButton = document.querySelector('button#fillCancelButton')
@@ -126,7 +125,7 @@ document.addEventListener('DOMContentLoaded', async function () {
   const button_docuChangelog = document.getElementById('button_openChangelog')
   const button_docuDatenschutz = document.getElementById('button_openDatenschutz')
   const button_openStore = document.getElementById('button_openStore')
-  const button_openLicense = document.getElementById('button_openLicense')
+  const button_openIssues = document.getElementById('button_openIssues')
   // module-platform element listeners
   const radio_bookingPlatforms = document.getElementsByName('booking-platform')
   const module_platform_element = document.getElementsByClassName('moduleItem-platform')
@@ -162,7 +161,7 @@ document.addEventListener('DOMContentLoaded', async function () {
   const privacyUrl = data_version.extension_privacy
   const readmeUrl = data_version.extension_readme
   const chromeStoreUrl = data_version.extension_chromestore
-  const licenseUrl = data_version.extension_license
+  const issuesUrl = data_version.extension_issues
   const version = data_version.extension_version
   const versionName = data_version.extension_version_name
   const buildVersion = data_version.extension_build
@@ -174,6 +173,7 @@ document.addEventListener('DOMContentLoaded', async function () {
   const supportedProfileTypes = data_version.supported_profile_types
   const updateTextOverview = data_version.extension_update_text_overview
   const updateTextDetails = data_version.extension_update_text_details
+  const updateClearModuleStorage = data_version.update_clear_moduleconfigurations
   // main action buttons functions
   function openConfigs() {
     if (window.configOpen) {
@@ -244,6 +244,9 @@ document.addEventListener('DOMContentLoaded', async function () {
     try {
       setStatusBarText(window.language.statusbartext_filterData)
       filterData = await filters(filter, clipboarsString)
+      if(!filterData || filterData === null || filterData.length < 1 ) {
+        throw('Unable to filter String. Filter data length: ',filterData.length)
+      }
       debugStick(filterData,"💽 Selected Filter-Module: " + filter)
     } catch (error) {
       console.error("❌ Unable to call bookingData: ", error + " | app");
@@ -259,6 +262,7 @@ document.addEventListener('DOMContentLoaded', async function () {
       debugStick(bookingPlatform,"🔘 Selected Platform-Module: ")
       setStatusBarText(window.language.statusbartext_passDataToPlatform)
       let bookEntries = await platforms(bookingPlatform, filterData, lstorage_cDetectionItems)
+      debugStick(bookEntries,"📝 Booking entries | ")
       let detailMessage = ''
       if (bookEntries.success) {
         debugStick(bookEntries,"✅ Booking process finished | ")
@@ -289,7 +293,11 @@ document.addEventListener('DOMContentLoaded', async function () {
       debugStick(error,"Booking Error-Feedback: ")
       lockActionButtons('false',fillButton)
       if(switch_showAllMessages.checked) {
-        message(true, error.errorstatus, window.language.error + ': ' + error.errorheadline, error.errortext || bookingPlatform)
+        // error message fallback, if wie get uncatched errors
+        let errorheadline = error.errorheadline ?? window.language.error_errorscript
+        let errortext = error.errortext ?? error + ' | '+ bookingPlatform
+        let errorstatus = error.errorstatus ?? 'error'
+        message(true, errorstatus, window.language.error + ': ' + errorheadline, errortext || bookingPlatform)
       } else {
         console.warn(consoleWarnMessage_showMessageTurnedOff)
       }
@@ -325,7 +333,7 @@ document.addEventListener('DOMContentLoaded', async function () {
     removeTabActiveClass()
     elem.classList.add('button-tab--active')
     let idName = elemId.split('-')[2]
-    let configWindowToShow = document.getElementById('config-win-'+idName)
+    let configWindowToShow = document.getElementById('configWindow_'+idName)
     configWindowToShow.classList.remove('dNone')
     if(idName === "general") {
       configurationsContainer.classList.add('configuration-container-first-tab-selected')
@@ -338,7 +346,7 @@ document.addEventListener('DOMContentLoaded', async function () {
   function timesheetFilterChange(e) {
     let timesheetFilterValue = e.target.value.split(filter_timesheetFilterPreValue)[1]
     localStorage.setItem('tc_c_filter', timesheetFilterValue)
-    markTabButtons('false','timesheets')
+    markTabButtons('false','filters')
     window.configUserChanges = true
   }
 
@@ -412,7 +420,7 @@ document.addEventListener('DOMContentLoaded', async function () {
 
   function showBuildVersion(e){
     if(e.shiftKey){
-      configItem_content_row_build_version.classList.remove('dNone')
+      configItem_content_row_buildversion.classList.remove('dNone')
     }
   }
 
@@ -445,33 +453,33 @@ document.addEventListener('DOMContentLoaded', async function () {
   }
 
   function moduleProTimeForceLatencyMode(e){
-    localStorage.setItem('tc_c_module_proTimeForceLatencyMode', e.target.checked)
+    localStorage.setItem(moduleStorage_preValueConfiguration+'_proTimeForceLatencyMode', e.target.checked)
     window.configUserChanges = true
   }
 
   function moduleProTimeUseLatencyMode(e){
-    localStorage.setItem('tc_c_module_proTimeUseLatencyMode', e.target.checked)
+    localStorage.setItem(moduleStorage_preValueConfiguration+'_proTimeUseLatencyMode', e.target.checked)
     window.configUserChanges = true
   }
 
   function moduleProTimeCheckUsePTTest(){
-    localStorage.setItem('tc_c_module_proTimeTest', moduleProTime_config_check_usePTTest.checked)
+    localStorage.setItem(moduleStorage_preValueConfiguration+'_proTimeTest', moduleProTime_config_check_usePTTest.checked)
     setModuleAmagProTimeTestStyle(moduleProTime_config_check_usePTTest.checked,...window.moduleGlobalArgs )
     window.configUserChanges = true
   }
 
   function moduleProTimeUseTicketNomberInText(e){
-    localStorage.setItem('tc_c_module_proTimeTicketNomberInText', e.target.checked)
+    localStorage.setItem(moduleStorage_preValueConfiguration+'_proTimeTicketNomberInText', e.target.checked)
     window.configUserChanges = true
   }
 
   function moduleProTimeUseMatchBookingDay(e) {
-    localStorage.setItem('tc_c_proTimeMatchBookingDay', e.target.checked)
+    localStorage.setItem(moduleStorage_preValueConfiguration+'_proTimeMatchBookingDay', e.target.checked)
     window.configUserChanges = true
   }
 
   function moduleProTimeUseAutoSelectDay(e) {
-    localStorage.setItem('tc_c_proTimeAutoSelectDay', e.target.checked)
+    localStorage.setItem(moduleStorage_preValueConfiguration+'_proTimeAutoSelectDay', e.target.checked)
     window.configUserChanges = true
   }
 
@@ -491,9 +499,9 @@ document.addEventListener('DOMContentLoaded', async function () {
   window.addEventListener("load", (event) => {
     // set global vars
     window.appVersionData = [{dokuUrl:dokuUrl,changelogUrl:changelogUrl,privacyUrl:privacyUrl,readmeUrl:readmeUrl,
-      chromeStoreUrl:chromeStoreUrl,licenseUrl:licenseUrl,version:version, versionName: versionName, buildVersion:buildVersion,
+      chromeStoreUrl:chromeStoreUrl,issuesUrl:issuesUrl,version:version, versionName: versionName, buildVersion:buildVersion,
       author:author,tester:tester,profileVersion:profileVersion,profileType: profileType,supportedProfileVersions:supportedProfileVersions,
-      supportedProfileTypes: supportedProfileTypes,updateTextOverview:updateTextOverview,updateTextDetails:updateTextDetails
+      supportedProfileTypes: supportedProfileTypes,updateTextOverview:updateTextOverview,updateTextDetails:updateTextDetails, updateClearModuleStorage: updateClearModuleStorage
     }]
     window.appGlobalArgs = [{elem_themeselect: themeSelect,configprofilename: configProfileName,link_csstheme: link_cssTheme,switch_showallmessages: switch_showAllMessages,
       switch_showStatusBar: switch_showStatusBar,elem_messagesection: elem_messageSection,messagesheadline: messagesHeadline, elem_configButton: configButton,elem_profilePictureUser: profilePictureUser, 
@@ -511,10 +519,10 @@ document.addEventListener('DOMContentLoaded', async function () {
     }
     // display version
     label_version.insertAdjacentHTML('beforeend', version)
-    label_version_name.insertAdjacentHTML('beforeend', versionName)
-    label_build_version.insertAdjacentHTML('beforeend', buildVersion)
-    label_extensionDevelop.insertAdjacentHTML('beforeend', author)
-    label_extensionCoDevelop.insertAdjacentHTML('beforeend', tester)
+    label_versionname.insertAdjacentHTML('beforeend', versionName)
+    label_buildversion.insertAdjacentHTML('beforeend', buildVersion)
+    label_extensiondevelop.insertAdjacentHTML('beforeend', author)
+    label_extensioncodevelop.insertAdjacentHTML('beforeend', tester)
     // add aceppted profile formats
     button_importConfigs.setAttribute('accept',profileType+", "+supportedProfileTypes.toString())
     // main action buttons listener
@@ -522,7 +530,7 @@ document.addEventListener('DOMContentLoaded', async function () {
     fillCancelButton.addEventListener('click', cancelPasteData)
     configButton.addEventListener('click', openConfigs)
     button_clearAllMessages.addEventListener('click', clearAllMessages)
-    buttonBackToMain.addEventListener('click', openConfigs)
+    button_backToMain.addEventListener('click', openConfigs)
     // configuration tabs listener
     buttonTab_General.addEventListener('click', () => switchConfigTab(buttonTab_General.id,buttonTab_General))
     buttonTab_Projects.addEventListener('click', () => switchConfigTab(buttonTab_Projects.id,buttonTab_Projects))
@@ -538,10 +546,10 @@ document.addEventListener('DOMContentLoaded', async function () {
     button_docuChangelog.addEventListener('click', () => window.open(changelogUrl))
     button_docuReadme.addEventListener('click', () => window.open(readmeUrl))
     button_openStore.addEventListener('click', () => window.open(chromeStoreUrl))
-    button_openLicense.addEventListener('click', () => window.open(licenseUrl))
+    button_openIssues.addEventListener('click', () => window.open(issuesUrl))
     // other
-    label_version_name.addEventListener('click', (e) => showBuildVersion(e))
-    label_extensionDevelop.addEventListener('click', (e) => setCreatorStorage(e))
+    label_versionname.addEventListener('click', (e) => showBuildVersion(e))
+    label_extensiondevelop.addEventListener('click', (e) => setCreatorStorage(e))
     profileOptionsSelect.addEventListener('change', selectProfileOption)
     //theme select
     themeSelect.addEventListener('change', switchTheme)
@@ -559,8 +567,16 @@ document.addEventListener('DOMContentLoaded', async function () {
     }
     try{
       // load needed app functions
-      profileManager(...window.appGlobalArgs,...appVersionData,...window.moduleGlobalArgs)
-      appStorage(...window.appGlobalArgs,...appVersionData,...window.moduleGlobalArgs)
+      try {
+        profileManager(...window.appGlobalArgs,...appVersionData,...window.moduleGlobalArgs)
+      } catch(e){
+        throw new Error('ProfileManager fehlerhaft | ',e)
+      }
+      try {
+        appStorage(...window.appGlobalArgs,...appVersionData,...window.moduleGlobalArgs)
+      } catch(e){
+        throw 'AppStorage fehlerhaft | '+e
+      }
       xmasModule()
       hlweenModule()
       // reset restart count