+++ src/components/content/firstStartDisplay/contentHtmls/firstStartDisplayHtml.js
@@ -74,7 +74,7 @@
             <div id="startScreenMain">
                 <div class="configItem">
                     <div class="configItem-content" id="startScreenContent">
-
+                        <!-- Dynamic Content here-->
                     </div>
                 </div>
             </div>